ch5(fix): Page buffering now accounts for partial pages
Previously, partial pages would error due to not meeting the check that
the number of bytes read must be the same as a pages size (4096) which
was in place to detect fread() errors. Now, if the read number of bytes
is a multiple of a row (minus padding) then page loading works as
expected and only load the partial page.