ch6: minisql no longer opens files in append mode
sqlmini was incorrectly opening db files in append mode causing all
writes to be added to the end of the file (appended, duh) and thus
causing data to be duplicated instead of over written. Instead files are
now opened in read+ mode and if fopen() fails due to the file not
existing, attempts to open a new file in write+ mode such to create the
file. If this fails, program hard fails and exists.