ch6: Memory footprint of internal structs compacted
Reverted to using uint32_t instead of uint64_t for most of the internal
unsigned int representations. uint64_t was chosen earlier due to
misunderstanding of a warning around the use of fseek() and its offset
parameter.
The smaller int type size also reducing the footprint of pages when
stored in memory and on disk as the larger range will not be necessary.